requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
Pillow==10.2.0